[REFACTORING] - Skiller ut oppføling.
Oppfølging virker å ikke nødvendigvis være et eget subdomene, men det er ihveftfall en integrasjon mot eksternt grensesnitt.
Da er det ryddig å tydeliggjøre dette, og ikke blande denne logikken sammen med ren domenelogikk, eller andre services.
