+++ src/main/java/no/nav/fo/veilarbregistrering/oppfolging/adapter/AktiverBrukerData.java
@@ -1,9 +1,10 @@
-package no.nav.fo.veilarbregistrering.domain;
+package no.nav.fo.veilarbregistrering.oppfolging.adapter;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
+import no.nav.fo.veilarbregistrering.domain.Innsatsgruppe;
 
 @Data
 @NoArgsConstructor
