+++ src/main/java/no/nav/fo/veilarbregistrering/oppfolging/adapter/SykmeldtBrukerType.java
@@ -1,4 +1,4 @@
-package no.nav.fo.veilarbregistrering.domain;
+package no.nav.fo.veilarbregistrering.oppfolging.adapter;
 
 public enum SykmeldtBrukerType {
     SKAL_TIL_NY_ARBEIDSGIVER,
